app/test: append 'allow' parameters to secondary processes
The 'allow' parameters are not passed to the secondary process, and all
devices will be loaded by default. When the primary process specifies
the 'allow' parameters and does not specify all devices that use the
vfio-pci driver, the secondary process will not load the devices as
expected, which will return incorrect test results.

This patch fixes this issue by appending the 'allow' parameters to the
secondary process.

Fixes: 086eb64db39e ("test/pdump: add unit test for pdump library")
Fixes: 148f963fb532 ("xen: core library changes")
Fixes: af75078fece3 ("first public release")
Fixes: b8d5e544e73e ("test: add procfs error message for multi-process launch")
Cc: stable@dpdk.org

Signed-off-by: Mingjin Ye <mingjinx.ye@intel.com>